Code now functional, though config panel is still a stub. Will complete soon.
