+++ settings.gradle
@@ -16,4 +16,4 @@ include 'api'
 include 'services:webservice'
 */
 
-rootProject.name = 'LightSpawnCBBase'
+rootProject.name = 'spawndetector'
